Cleaned up testing and added more Test Cases.
FileGen Test:
1. Resolved issue with tester class not passing. Issue was due to
currentIndent not being reset for each test. It is now reset in
openWriter().
2. Added more test cases, included tests for the currentIndent counter.

Parser Test:
Added testing for noteType(). Found errors with the beatNotes that are
fractions, as it was founding it to 0.0. May want to accept input as a
string, and convert to double if needed.

Parser:
Changed fractions in noteType from 1/# to 1.0/# to use double division.
Method may need to be adapted, as it does not work if fraction is
entered as input.